systems/lead-gen-machine/requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
anthropic>=0.40.0
firecrawl-py>=1.0.0
